Suggestion from Sondre regarding making sure that internal_dict always is initialized. And splitted tests into multiple files
