+++ tests/pyaaas/models/test_privacy_models_ldiversity.py
@@ -1,21 +1,8 @@
 from pyaaas.models.privacy_models import LDiversityDistinct, LDiversityGrassbergerEntropy, LDiversityShannonEntropy, LDiversityRecursive
-from pyaaas.aaas import AaaS
 import unittest
 
+
 class PrivacyModelsLDiversityTest(unittest.TestCase):
-
-    def test_privacyModelIterator(self):
-        self.privacymodel = LDiversityDistinct(l=3, column_name="column")
-        self.assertEqual(list(iter(self.privacymodel))[1], "column_name", "Iterator does not create correct list of values")
-
-    def test_lenght(self):
-        self.privacymodel = LDiversityDistinct(l=3, column_name="column")
-        self.assertEqual(len(self.privacymodel), 2)
-
-    def test_message(self):
-        self.privacymodel = LDiversityDistinct(l=3, column_name="column")
-        self.assertEqual(str(self.privacymodel), "LDiversityDistinct(l=3, column_name=column)")
-
     def test_set_data_distinct(self):
         self.distict = LDiversityDistinct(l=3, column_name="sensitive_column_name")
         self.assertEqual(self.distict._anonymity_name, "LDIVERSITY_DISTINCT", "Recursive model name not correct")
@@ -23,7 +10,6 @@ class PrivacyModelsLDiversityTest(unittest.TestCase):
     def test_set_data_grassberger(self):
         self.grassberger = LDiversityGrassbergerEntropy(l=3, column_name="sensitive_column_name")
         self.assertEqual(self.grassberger._anonymity_name, "LDIVERSITY_GRASSBERGERENTROPY", "Grassberger model name not correct")
-
 
     def test_set_data_shannonentropy(self):
         self.shannon_entropy = LDiversityShannonEntropy(l=3, column_name="sensitive_column_name")
